pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel", "pybind11>=2.10.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kompass_core"
version = "0.0.1"
description = "Kompass Navigation Algorithms"
authors = [{ name = "Automatika Robotics", email = "contact@automatikarobotics.com" }]
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "numpy==1.26.4",
    "numpy-quaternion==2023.0.3",
    "python-fcl@git+https://github.com/automatika-robotics/python-fcl",
    "pybind11",
    "omegaconf",
    "attrs==23.2.0",
    "numba",
    "pyyaml",
    "scipy",
    "matplotlib",
    "pandas",
]

[project.optional-dependencies]
dev = ["pip-tools", "pytest==8.0.0", "pre-commit"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
testpaths = [
    "tests"
]
python_files = '*pytest.py'

[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
ignore-magic = false
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = true
ignore-nested-functions = false
ignore-nested-classes = true
ignore-setters = false
exclude = ["setup.py", "docs", "build", "tests"]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*"]
quiet = false
whitelist-regex = []
color = true
generate-badge = "."
badge-format = "svg"

[tool.ruff]
extend-exclude = [".mypy_cache", ".tox", ".venv", "buck-out", "build", ".pytest_cache"]
fix = true
line-length = 88
preview = true
[tool.ruff.lint]
ignore = ["E203", "E266", "E501", "F403", "F401"]
select = ["B","C","E","F","W","B9"]
[tool.ruff.lint.mccabe]
max-complexity = 10
